requirements.txt
# 智能象棋机器人项目依赖包

# 深度学习和视觉处理
# torch==2.0.0  # ARM64版本需要单独安装
# torchvision==0.15.0  # ARM64版本需要单独安装
opencv-python==4.8.0.74
numpy==1.24.3
pillow==10.0.0
matplotlib==3.7.2
scikit-image==0.21.0

# 机器人控制
pyserial==3.5
scipy==1.10.1
pandas==2.0.3

# 象棋AI
stockfish==3.28.0
python-chess==1.999

# Web框架和API
fastapi==0.104.0
uvicorn==0.23.2
websockets==11.0.3
pydantic==2.4.2
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# 数据库驱动
motor==3.3.1
redis==5.0.0
pymongo==4.5.0

# 测试框架
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-mock==3.11.1

# 监控和日志
prometheus-client==0.17.1
structlog==23.1.0
psutil==5.9.5
python-json-logger==2.0.7

# 异步支持
aioredis==2.0.1
asyncio-mqtt==0.13.0

# 认证和安全
pyjwt==2.8.0
bcrypt==4.0.1
cryptography==41.0.7
python-multipart==0.0.6
email-validator==2.1.0
aiohttp==3.8.6

# 性能监控
pynvml==11.5.0